Fix get_commitment_proof

+++ libzeropool-rs/src/merkle.rs
@@ -261,7 +261,7 @@ impl<D: KeyValueDB, P: PoolParams> MerkleTree<D, P> {
         &self,
         index: u64,
     ) -> Option<MerkleProof<P::Fr, { constants::HEIGHT - constants::OUTLOG }>> {
-        let key = Self::node_key(constants::OUTLOG, index);
+        let key = Self::node_key(constants::OUTLOG as u32, index);
         let node_present = self.db.get(0, &key).map_or(false, |value| value.is_some());
         if !node_present {
             return None;
